requirements.txt
Pillow==9.2.0
pandas==1.4.3